space/requirements.txt
flair==0.8.0.post1
st-annotated-text